+++ Assets/Scripts/UIManager.cs
@@ -37,7 +37,7 @@ public class UIManager : MonoBehaviour {
 	public AudioSource audioSource;
 
 
-	public void AddScore(float type, ObjectManager.MergeLevel mergeLevel = ObjectManager.MergeLevel.one) {
+	public void AddScore(int type, ObjectManager.MergeLevel mergeLevel = ObjectManager.MergeLevel.one) {
 		ScoreManager.init.AddScore(type, mergeLevel);
     }
